- Move utility methods from proton::attribute to search::attribute. - Compute isUpdateableInmemoryOnly once.
