+++ searchcommon/src/vespa/searchcommon/attribute/CMakeLists.txt
@@ -1,6 +1,7 @@
 # Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 vespa_add_library(searchcommon_searchcommon_attribute OBJECT
     SOURCES
+    attribute_utils.cpp
     basictype.cpp
     collectiontype.cpp
     config.cpp
