+++ searchlib/src/vespa/searchlib/attribute/attributevector.h
@@ -397,6 +397,7 @@ public:
     CollectionType getInternalCollectionType() const { return _config.collectionType(); }
     const BaseName & getBaseFileName() const { return _baseFileName; }
     void setBaseFileName(vespalib::stringref name) { _baseFileName = name; }
+    bool isUpdateableInMemoryOnly() const { return _isUpdateableInMemoryOnly; }
 
     const vespalib::string & getName() const override final { return _baseFileName.getAttributeName(); }
 
@@ -575,7 +576,7 @@ private:
     BaseName                              _baseFileName;
     Config                                _config;
     std::shared_ptr<attribute::Interlock> _interlock;
-    mutable std::shared_mutex       _enumLock;
+    mutable std::shared_mutex             _enumLock;
     GenerationHandler                     _genHandler;
     GenerationHolder                      _genHolder;
     Status                                _status;
@@ -587,6 +588,7 @@ private:
     uint64_t                              _compactLidSpaceGeneration;
     bool                                  _hasEnum;
     bool                                  _loaded;
+    bool                                  _isUpdateableInMemoryOnly;
     vespalib::steady_time                 _nextStatUpdateTime;
 
 ////// Locking strategy interface. only available from the Guards.
